requirements.txt
fastapi
uvicorn
scikit-learn
numpy
psycopg2-binary
